refactor(parser): move Diagnostic AST node
